+++ Source/Mods/HitmanPhone/src/Mod.cs
@@ -18,7 +18,7 @@ using S1NPC = ScheduleOne.NPCs.NPC;
 using S1NPCHealth = ScheduleOne.NPCs.NPCHealth;
 #endif
 
-[assembly: MelonInfo(typeof(HitmanPhone.Mod), "HitmanPhone", "0.2.8", "Dominik")]
+[assembly: MelonInfo(typeof(HitmanPhone.Mod), "HitmanPhone", "0.2.9", "Dominik")]
 [assembly: MelonGame("TVGS", "Schedule I")]
 
 namespace HitmanPhone;